requirements.txt
python-dotenv
pandas
numpy
setuptools
pymongo[srv]==3.12
pymongo
python-dotenv
certifi
scikit-learn
scipy
dill
pyyaml
##-e .